+++ ModelMID/DB/AddWeight.cs
@@ -6,7 +6,7 @@ namespace ModelMID.DB
 {
     public class AddWeight
     {
-        public int CodeWares { get; set; }
+        public long CodeWares { get; set; }
         public int CodeUnit { get; set; }
         public decimal Weight { get; set; }
         public bool IsManual { get; set; }
