+++ ModelMID/DB/AdditionUnit.cs
@@ -6,7 +6,7 @@ namespace ModelMID
 {
     public class AdditionUnit
     {
-        public int CodeWares { get; set; }
+        public long CodeWares { get; set; }
         public int CodeUnit { get; set; }
         public decimal Coefficient { get; set; }
         public decimal Weight { get; set; }
