+++ ModelMID/DB/BarCode.cs
@@ -6,7 +6,7 @@ namespace ModelMID
 {
     public class Barcode
     {
-        public int CodeWares { get; set; }
+        public long CodeWares { get; set; }
         public int CodeUnit { get; set; }
         public string BarCode { get; set; }
         public decimal Coefficient { get; set; }
