+++ ModelMID/DB/FiscalArticle.cs
@@ -12,13 +12,13 @@ namespace ModelMID.DB
         /// <summary>
         /// Код товару
         /// </summary>
-        public int CodeWares { get; set; }
+        public long CodeWares { get; set; }
         /// <summary>
         /// Назва товару
         /// </summary>
         public string NameWares { get; set; }
 
-        public int PLU { get; set; }
+        public long PLU { get; set; }
         public decimal Price { get; set; }        
     }
 }
