+++ ModelMID/DB/ReceiptWaresLink.cs
@@ -6,7 +6,7 @@ namespace ModelMID.DB
 {   
     public class ReceiptWaresLink:IdReceiptWares
     {        
-        public int CodeWaresTo { get; set; }        
+        public Int64 CodeWaresTo { get; set; }        
         public decimal Sort { get; set; }
         public int  Quantity { get; set; }
         public string NameWares { get; set; }
